pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "word-template-generator"
version = "0.1.0"
description = "Local-first DOCX generator for Word templates and markdown acts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "docxtpl",
  "typer",
  "PyYAML",
  "fastapi",
  "uvicorn",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
word-gen = "word_template_generator.cli:main"

